Move helper functions to `helpers_linux.go`
There were a variety of helper functions defined in
`container_manager_linux.go`, which could just as easily be defined in
`helpers_linux.go`. Move the helper functions to `helpers_linux.go`
so `container_manager_linux.go` is more focused on the logic of managing
containers, and less focused on interacting with the OS (i.e. making
calculations about pids, etc...)
